[mmptest] Port link-httpclient-async to an NUnit LinkAll+LinkSdk test. Partial fix for #4975.
Fixes part of https://github.com/xamarin/xamarin-macios/issues/4975.
